feature/no-animation: moved grid non-hook functions to lib
